@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "^4.6.2",
+    "@ant-design/icons": "^4.7.0",
     "@craco/craco": "^6.1.1",
     "@crowdin/cli": "^3.6.4",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "antd": "^4.15.5",
+    "antd": "^4.22.8",
     "codemirror": "^5.61.1",
     "copy-to-clipboard": "^3.3.1",
-    "core-js": "^3.21.1",
+    "core-js": "^3.25.0",
     "craco-less": "^1.17.1",
     "eslint-plugin-unused-imports": "^2.0.0",
     "file-saver": "^2.0.5",
@@ -20,17 +20,17 @@
     "i18next": "^19.8.9",
     "moment": "^2.29.1",
     "qs": "^6.10.2",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-app-polyfill": "^3.0.0",
     "react-codemirror2": "^7.2.1",
     "react-cropper": "^2.1.7",
-    "react-device-detect": "^1.14.0",
-    "react-dom": "^17.0.2",
+    "react-device-detect": "^2.2.2",
+    "react-dom": "^18.2.0",
     "react-github-corner": "^2.5.0",
     "react-helmet": "^6.1.0",
-    "react-highlight-words": "^0.17.0",
+    "react-highlight-words": "^0.18.0",
     "react-i18next": "^11.8.7",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^5.3.3",
     "react-scripts": "4.0.3",
     "react-social-login-buttons": "^3.4.0"
   },
